:art: 테이블 아이디 BIGINT로 수정

+++ zzic-back/src/main/java/point/zzicback/mapper/TodoMapper.java
@@ -17,13 +17,13 @@ public interface TodoMapper {
      * 아래는 보편적인 MyBatis Mapper 메소드들입니다.
      * 이미 기능이 구현되어 있으며, `TodoMapper.xml` 파일에 정의되어 있습니다.
      */
-    int deleteByPrimaryKey(Integer id); // JPA의 deleteById와 같은 역할
+    int deleteByPrimaryKey(Long id); // JPA의 deleteById와 같은 역할
 
     int insert(Todo record); // JPA의 save와 같은 역할
 
     int insertSelective(Todo record); // JPA의 save와 같은 역할
 
-    Todo selectByPrimaryKey(Integer id);  // JPA의 findById와 같은 역할
+    Todo selectByPrimaryKey(Long id);  // JPA의 findById와 같은 역할
 
     int updateByPrimaryKeySelective(Todo record);  // JPA의 save와 같은 역할
